poker_mentor/requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
python-dotenv==1.0.0